Refactored our entire API layer
The purpose was really only to remove the "FindAnyObjectByType" call.
However, as I did, I refactored the entire class structure.
- LootLockerServerManager was only used in LootLockerBaseServerAPI and
  LootLockerGameServerAPI to get access to StartCoroutine. That was it's
only use. So I removed it and baked that functionality straight into
those classes.
- LootLockerGameServerAPI was a sub class of LootLockerBaseServerAPI in
  theory. But in reality they were so tightly coupled that it was a
complete fake inheritance. Additionally, the only reason for the
sub-class was a premature optimization to support f.ex. Server API and
Admin API through different subclasses. But that functionality was
instead built in Base through the SwitchURL method.
- Additionally, I changed the interfaces to decouple the stack. Instead
  of the user calling SwitchURL, getting the instance, and more, I baked
all that into the serverapi class and just provide a static "Send(...)"
method.